% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{get_E_step}
\alias{get_E_step}
\title{Run a Particle Filter on a POD based on a continuous LV model}
\usage{
get_E_step(obs_, obsTimes_, myParams, testedParams, n_part, n_dens_samp)
}
\arguments{
\item{obsTimes_}{vector of ordered obs times}

\item{n_part}{number of particles}

\item{n_dens_samp}{number of sampling for density}

\item{a1}{a2 Prey predatorsparameters}

\item{obs}{matrix of observations}

\item{cov}{qs RWC some parameters}
}
\value{
a list with particles and weights
}
\description{
Run a Particle Filter on a POD based on a continuous LV model
}
