\name{LotkaVoltR-package}
\alias{LotkaVoltR-package}
\alias{LotkaVoltR}
\docType{package}
\title{
\packageTitle{LotkaVoltR}
}
\description{
\packageDescription{LotkaVoltR}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{LotkaVoltR}
\packageIndices{LotkaVoltR}
To come
}
\author{
\packageAuthor{LotkaVoltR}

Maintainer: \packageMaintainer{LotkaVoltR}
}
\references{
To come
}
\keyword{ LotkaVolterra }
\seealso{
Optional links to other man pages, e.g.
}
\examples{
To come
}
