% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LVPOD.R
\name{LV_density}
\alias{LV_density}
\title{Unbiased density for a given Lotka Volterra between two points}
\usage{
LV_density(LV_model, x0, xT, time_lag, n_samples = 50)
}
\arguments{
\item{LV_model}{an object of class LVModel}

\item{x0}{starting 2d point}

\item{xT}{ending 2d point}

\item{time_lag}{time_lag between observations}
}
\value{
a (signed) estimate of the density
}
\description{
function performing CIS
}
