% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LVPOD.R
\name{LV_simulate}
\alias{LV_simulate}
\title{simulate}
\usage{
LV_simulate(LV_model, times = seq(0, 10, length.out = 2 * 10^4),
  selection = T)
}
\arguments{
\item{LV_model}{an object of class LVModel}

\item{selection}{selected points, default to all
is valid}

\item{simTimes}{simulation times. Should be with low increment so that the scheme}
}
\value{
a LVModel
}
\description{
function simulating from a Lotka Volterra model
}
