% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProposalLVModel.R
\name{PLV_getW0}
\alias{PLV_getW0}
\title{GetWeights of a Particle}
\usage{
PLV_getW0(PLV, part0, Obs)
}
\arguments{
\item{PLV}{PLV model}

\item{Obs}{First observation}
}
\value{
a set of weights
}
\description{
function that creates a LVPOD instance
}
