% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProposalLVModel.R
\name{PLV_simX0}
\alias{PLV_simX0}
\title{Simulate initial particle}
\usage{
PLV_simX0(PLV, Obs, N)
}
\arguments{
\item{PLV}{PLV model}

\item{Obs}{First observation}

\item{N}{Number of particles}
}
\value{
a Proposal LVPOD
}
\description{
function that creates a LVPOD instance
}
