% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/POLV_model.R
\name{POLV_density}
\alias{POLV_density}
\title{Unbiased density for a given Lotka Volterra between two points}
\usage{
POLV_density(POLV_model, x0, xT, time_lag, n_samples = 50)
}
\arguments{
\item{POLV_model}{an object of class LVModel}

\item{x0}{starting 2d point}

\item{xT}{ending 2d point}

\item{time_lag}{time_lag between observations}
}
\value{
a (signed) estimate of the density
}
\description{
function performing CIS
}
