% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/POLV_model.R
\name{POLV_simulate}
\alias{POLV_simulate}
\title{simulate}
\usage{
POLV_simulate(POLV_model, times = seq(0, 10, length.out = 2 * 10^4),
  selection = T)
}
\arguments{
\item{POLV_model}{an object of class POLVModel}

\item{selection}{selected points, default to all
is valid}

\item{simTimes}{simulation times. Should be with low increment so that the scheme}
}
\value{
a LVModel
}
\description{
function simulating from a Lotka Volterra model
}
