% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{ProposalLVModel}
\alias{ProposalLVModel}
\title{ProposalModel for LVPOD}
\description{
Proposoal model with optimal filter
POD with the bivariate SDE LV model
and the observation model Y_t = X_t * exp(N(0, Sigma_obs))
}
