% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{runLV_PF}
\alias{runLV_PF}
\title{Run a Particle Filter on a POD based on a continuous LV model}
\usage{
runLV_PF(obs_, obsTimes_, myParams, n_part, n_dens_samp)
}
\arguments{
\item{obsTimes_}{vector of ordered obs times}

\item{n_part}{number of particles}

\item{n_dens_samp}{number of sampling for density}

\item{a1}{a2 Prey predatorsparameters}

\item{obs}{matrix of observations}

\item{cov}{qs RWC some parameters}
}
\value{
a list with particles and weights
}
\description{
Run a Particle Filter on a POD based on a continuous LV model
}
